requirements.txt
aioredis
minio
pandas
pycaret
redis
tensorflow
wheel
surrealdb
weaviate-client
fireworks-ai
langchain
langchain-community
langchain_fireworks
langchain_weaviate
